Fixing the bad selector caching of permissions data 🤞
