+++ frontend/src/components/AddDevice.tsx
@@ -3,8 +3,7 @@ import { Box, Typography } from '@mui/material'
 import { platforms, IPlatform } from '../platforms'
 import { useSelector, useDispatch } from 'react-redux'
 import { ApplicationState, Dispatch } from '../store'
-import { selectPermissions } from '../models/organization'
-import { selectOrganization } from '../selectors/organizations'
+import { selectPermissions, selectOrganization } from '../selectors/organizations'
 import { DataCopy } from '../components/DataCopy'
 import { Notice } from '../components/Notice'
 import { Link } from '../components/Link'
