+++ frontend/src/models/networks.ts
@@ -2,10 +2,9 @@ import structuredClone from '@ungap/structured-clone'
 import { createModel } from '@rematch/core'
 import { isPortal } from '../services/Browser'
 import { ApplicationState } from '../store'
-import { getActiveUser } from './accounts'
-import { getActiveAccountId } from '../selectors/accounts'
 import { selectNetworks } from '../selectors/networks'
 import { selectConnection } from '../selectors/connections'
+import { getActiveAccountId, getActiveUser } from '../selectors/accounts'
 import { IOrganizationState, canMemberView, canViewByTags, canRoleView } from '../models/organization'
 import { selectById } from '../selectors/devices'
 import {
