+++ frontend/src/pages/NetworksPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Typography, Divider } from '@mui/material'
 import { selectNetworks } from '../selectors/networks'
-import { selectPermissions } from '../models/organization'
+import { selectPermissions } from '../selectors/organizations'
 import { ApplicationState, Dispatch } from '../store'
 import { useSelector, useDispatch } from 'react-redux'
 import { getDeviceModel } from '../selectors/devices'
