+++ frontend/src/pages/OrganizationGuestPage.tsx
@@ -3,8 +3,8 @@ import { getAllDevices } from '../selectors/devices'
 import { REGEX_LAST_PATH } from '../shared/constants'
 import { useSelector, useDispatch } from 'react-redux'
 import { useParams, useHistory, useLocation } from 'react-router-dom'
+import { selectOrganization, selectRemoteitLicense } from '../selectors/organizations'
 import { getFreeLicenses } from '../models/plans'
-import { selectRemoteitLicense } from '../selectors/plans'
 import { selectAccessibleNetworks } from '../models/networks'
 import { selectNetworks } from '../selectors/networks'
 import { ApplicationState, Dispatch } from '../store'
@@ -13,7 +13,6 @@ import { ListItemLocation } from '../components/ListItemLocation'
 import { TargetPlatform } from '../components/TargetPlatform'
 import { ShareDetails } from '../components/ShareDetails'
 import { RoleAccessCounts } from '../components/RoleAccessCounts'
-import { selectOrganization } from '../selectors/organizations'
 import { LinearProgress } from '../components/LinearProgress'
 import { ConfirmButton } from '../buttons/ConfirmButton'
 import { LicenseSelect } from '../components/LicenseSelect'
