+++ frontend/src/pages/OrganizationGuestsPage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { ApplicationState } from '../store'
 import { useSelector } from 'react-redux'
 import { Typography } from '@mui/material'
-import { selectPermissions } from '../models/organization'
+import { selectPermissions } from '../selectors/organizations'
 import { OrganizationGuestList } from '../components/OrganizationGuestList'
 import { Container } from '../components/Container'
 import { Gutters } from '../components/Gutters'
