+++ frontend/src/pages/OrganizationMembersPage.tsx
@@ -1,11 +1,8 @@
 import React from 'react'
 import { ApplicationState } from '../store'
 import { useSelector } from 'react-redux'
-import { selectOwner } from '../models/organization'
-import { selectOrganization } from '../selectors/organizations'
 import { Typography, List } from '@mui/material'
-import { selectPermissions } from '../models/organization'
-import { selectRemoteitLicense } from '../selectors/plans'
+import { selectRemoteitLicense, selectPermissions, selectOrganization, selectOwner } from '../selectors/organizations'
 import { OrganizationMemberList } from '../components/OrganizationMemberList'
 import { LicensingNoticeDisplay } from '../components/LicensingNoticeDisplay'
 import { SeatsSetting } from '../components/SeatsSetting'
