+++ frontend/src/pages/OrganizationRolesPage.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { Redirect, useHistory } from 'react-router-dom'
 import { List, ListItemSecondaryAction, Typography, Chip, Box, Button } from '@mui/material'
-import { selectLimitsLookup, selectOrganization } from '../selectors/organizations'
-import { selectPermissions } from '../models/organization'
+import { selectPermissions, selectLimitsLookup, selectOrganization } from '../selectors/organizations'
 import { ApplicationState } from '../store'
 import { ListItemLocation } from '../components/ListItemLocation'
 import { useSelector } from 'react-redux'
